+++ src/main/scala/tile/FPU.scala
@@ -874,9 +874,3 @@ class FPU(cfg: FPUParams)(implicit p: Parameters) extends FPUModule()(p) {
     req
   }
 }
-
-/** Mix-in for constructing tiles that may have an FPU external to the core pipeline */
-trait CanHaveSharedFPUModule[+L <: BaseTile] { this: BaseTileModule[L] =>
-  val fpuOpt = outer.tileParams.core.fpu.map(params => Module(new FPU(params)(outer.p)))
-  // TODO fpArb could go here instead of inside LegacyRoccComplex
-}
